+++ next.config.js
@@ -32,7 +32,7 @@ const nextConfig = {
         pathname: "/**",
       },
     ],
-    deviceSizes: [640, 750, 828],
+    deviceSizes: [640],
     imageSizes: [16, 32, 48, 64, 96, 128, 256, 384, 640, 750],
   },
   transpilePackages: ["notion-client", "notion-types", "react-notion-x"],
