--- src/app/api/tags/route.ts
@@ -1,16 +1,0 @@
-import { NextResponse } from "next/server";
-import { getUniqueTags } from "@/shared/lib/api/notion";
-import { PostType } from "@/shared/types/notion";
-
-export async function GET() {
-  try {
-    const tags = await getUniqueTags(PostType.blog);
-    return NextResponse.json(tags);
-  } catch (error) {
-    console.error("Error fetching tags:", error);
-    return NextResponse.json(
-      { error: "Failed to fetch tags" },
-      { status: 500 }
-    );
-  }
-}
