+++ src/app/blog/page.tsx
@@ -1,5 +1,4 @@
-import { getPosts, getUniqueTags } from "@/shared/lib/api/notion";
-import { PostType } from "@/shared/types/notion";
+import { getPosts, getUniqueTags, PostType } from "@/domain/posts";
 import BlogScreen from "@/screens/blog/BlogScreen";
 
 export default async function BlogPage() {
