+++ src/domain/index.ts
@@ -1,0 +1,2 @@
+export * from "./posts";
+export * from "./notion";
