+++ src/domain/notion/api/index.ts
@@ -1,0 +1,1 @@
+export * from "./getRecordMap";
