+++ src/domain/posts/api/index.ts
@@ -1,0 +1,3 @@
+export * from "./getPosts";
+export * from "./getPageIdBySlug";
+export * from "./getUniqueTags";
