+++ src/domain/posts/types.ts
