--- src/shared/lib/api/notion/index.ts
@@ -1,2 +1,0 @@
-export * from "./notionhqClient";
-export * from "./notionClient";
