+++ src/shared/lib/hooks/index.ts
@@ -1,3 +1,1 @@
-export { default as useFetch } from "./useFetch";
 export { default as useHeadings } from "./useHeadings";
-export { default as useNotionRecord } from "./useNotionRecord";
