--- src/shared/lib/hooks/useFetch.ts
@@ -1,27 +1,0 @@
-import { useEffect, useState } from "react";
-
-export default function useFetch<T>(url: string, initialData: T) {
-  const [data, setData] = useState<T>(initialData);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<Error | null>(null);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setLoading(true);
-      try {
-        const response = await fetch(url);
-        if (!response.ok) throw new Error(`Failed to fetch: ${url}`);
-        const result = await response.json();
-        setData(result);
-      } catch (err) {
-        console.error("Error fetching data:", err);
-        setError(err as Error);
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchData();
-  }, [url]);
-
-  return { data, loading, error };
-}
